+++ src/main/java/com/barco/service1/controller/dashboard/StockTradeDashboardController.java
@@ -1,0 +1,20 @@
+package com.barco.service1.controller.dashboard;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * Api use to perform crud operation
+ * @author Nabeel Ahmed
+ */
+@RestController
+@CrossOrigin(origins="*")
+@RequestMapping(value="/stockTradeDashboard.json")
+public class StockTradeDashboardController {
+
+    private Logger logger = LoggerFactory.getLogger(StockTradeDashboardController.class);
+
+}
